pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kub-health"
version = "0.1.0"
description = "AI-powered Kubernetes cluster health screening tool"
requires-python = ">=3.10"
dependencies = [
    "kubernetes>=28.1.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
kub-health = "kub_health.cli:main"

[tool.setuptools.packages.find]
include = ["kub_health*"]

[tool.ruff]
line-length = 100
target-version = "py310"
